requirements.txt
fastapi==0.117.0
uvicorn[standard]==0.24.0
SQLAlchemy[asyncio]==2.0.20
asyncpg==0.31.0
alembic==1.12.0
python-jose[cryptography]==3.2.0
passlib[bcrypt]==1.7.4
pydantic==2.8.0
pydantic-settings==2.2.1
python-dotenv==1.0.0
slowapi==0.1.9
httpx==0.28.1
pytest==7.4.2
pytest-asyncio==0.21.1
python-multipart==0.0.6
